pyproject.toml
[tool.poetry]
name = "tajana"
version = "0.1.0"
description = "test files"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.2.2"
vectorbt = "^0.26.2"
yfinance = "^0.2.41"
nbformat = "^5.10.4"
xbbg = "^0.7.7"
quantstats = "^0.0.62"
ipykernel = "^6.29.5"
plotly = "^5.23.0"
openpyxl = "^3.1.5"
jinja2 = "^3.1.4"
tsfresh = "^0.20.3"
setuptools = "^73.0.1"
dash = "^2.17.1"
jupyter-dash = "^0.4.2"
grid = "^0.7.1"
qgrid = "^1.3.1"
chart-studio = "^1.1.0"
nbconvert = {extras = ["webpdf"], version = "^7.16.4"}
notebook = "^7.2.2"
playwright = "^1.47.0"
sqlalchemy = "^2.0.25"
pyyaml = "^6.0.1"
jupyter = "^1.1.1"
pdblp = "^0.1.8"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
